feat(#540): move stack computation to a separate class MaxStack
