+++ src/main/java/org/eolang/jeo/representation/bytecode/BytecodeMethod.java
@@ -23,15 +23,9 @@
  */
 package org.eolang.jeo.representation.bytecode;
 
-import com.jcabi.log.Logger;
-import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Deque;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.TreeMap;
 import java.util.stream.Collectors;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
@@ -42,7 +36,6 @@ import org.eolang.jeo.representation.xmir.AllLabels;
 import org.objectweb.asm.Label;
 import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Opcodes;
-import org.objectweb.asm.Type;
 
 /**
  * Bytecode method.
@@ -427,80 +420,19 @@ public final class BytecodeMethod implements Testable {
         return this.directives(true);
     }
 
+    /**
+     * Compute max stack.
+     * @return Max stack.
+     */
     private int computeStack() {
-        Logger.info(this, "Computing stack for %s", this.properties);
-        int max = 0;
-        final Deque<Integer> worklist = new ArrayDeque<>(0);
-        final int length = this.instructions.size();
-        worklist.add(0);
-        Map<Integer, Integer> visited = new TreeMap<>();
-        this.tryblocks.stream()
-            .map(BytecodeTryCatchBlock.class::cast)
-            .map(BytecodeTryCatchBlock::handler)
-            .map(this::index)
-            .peek(ind -> visited.put(ind, 1))
-            .forEach(worklist::add);
-        while (!worklist.isEmpty()) {
-            int current = worklist.pop();
-            int stack = visited.get(current) == null ? 0 : visited.get(current);
-            while (current < length) {
-                BytecodeEntry entry = this.instructions.get(current);
-                stack += entry.stackImpact();
-                max = Math.max(max, stack);
-                final int finalStack = stack;
-                visited.compute(
-                    current, (k, v) -> v == null ? finalStack : Math.max(v, finalStack)
-                );
-                if (entry instanceof BytecodeInstruction) {
-                    final BytecodeInstruction var = BytecodeInstruction.class.cast(entry);
-                    if (var.isBranchInstruction()) {
-                        if (var.isSwitchInstruction()) {
-                            final List<Label> offsets = var.offsets();
-                            for (Label offset : offsets) {
-                                final int target = this.index(offset);
-                                if (visited.get(target) == null
-                                    || visited.get(target) < stack
-                                ) {
-                                    visited.put(target, stack);
-                                    worklist.add(target);
-                                }
-                            }
-                            break;
-                        } else if (var.isConditionalBranchInstruction()) {
-                            final int jump = this.index(var.offset());
-                            if (visited.get(jump) == null
-                                || visited.get(jump) < stack
-                            ) {
-                                visited.put(jump, stack);
-                                worklist.add(jump);
-
-                            }
-                            final int next = current + 1;
-                            if (visited.get(next) == null
-                                || visited.get(next) < stack
-                            ) {
-                                visited.put(next, stack);
-                                worklist.add(next);
-                            }
-                            break;
-                        } else if (var.isReturnInstruction()) {
-                            break;
-                        } else {
-                            final int jump = this.index(var.offset());
-                            if (visited.get(jump) == null
-                                || visited.get(jump) < stack
-                            ) {
-                                visited.put(jump, stack);
-                                worklist.add(jump);
-                            }
-                            break;
-                        }
-                    }
-                }
-                current++;
-            }
-        }
-        return max;
+        return new MaxStack(
+            this.properties,
+            this.instructions,
+            this.tryblocks.stream()
+                .filter(BytecodeTryCatchBlock.class::isInstance)
+                .map(BytecodeTryCatchBlock.class::cast)
+                .collect(Collectors.toList())
+        ).value();
     }
 
     /**
@@ -509,134 +441,14 @@ public final class BytecodeMethod implements Testable {
      */
     private int computeLocals() {
         return new MaxLocals(
-            this.properties, this.instructions,
+            this.properties,
+            this.instructions,
             this.tryblocks.stream()
                 .filter(BytecodeTryCatchBlock.class::isInstance)
                 .map(BytecodeTryCatchBlock.class::cast)
                 .collect(Collectors.toList())
         ).value();
-//        Logger.info(this, "Computing locals for %s", this.properties);
-//        Variables initial = new Variables();
-//        int first = 0;
-//        if (!this.properties.isStatic()) {
-//            initial.put(0, 1);
-//            first = 1;
-//        }
-//        final Type[] types = Type.getArgumentTypes(this.properties.descriptor());
-//        for (int index = 0; index < types.length; index++) {
-//            final Type type = types[index];
-//            initial.put(index * type.getSize() + first, type.getSize());
-//        }
-//        Map<Integer, Variables> all = new TreeMap<>();
-//        Map<Integer, Variables> worklist = new HashMap<>();
-//        worklist.put(0, initial);
-//        final int total = this.instructions.size();
-//        Variables currentVars;
-//        while (!worklist.isEmpty()) {
-//            final Map.Entry<Integer, Variables> curr = worklist.entrySet()
-//                .stream()
-//                .findFirst()
-//                .orElseThrow(() -> new IllegalStateException(""));
-//            int current = curr.getKey();
-//            worklist.remove(current);
-//            if (all.get(current) != null) {
-//                continue;
-//            }
-//            currentVars = new Variables(curr.getValue());
-//            while (current < total) {
-//                BytecodeEntry entry = this.instructions.get(current);
-//                if (entry instanceof BytecodeInstruction) {
-//                    final BytecodeInstruction var = BytecodeInstruction.class.cast(entry);
-//                    if (var.isBranchInstruction()) {
-//                        if (var.isSwitchInstruction()) {
-//                            final List<Label> offsets = var.offsets();
-//                            for (Label offset : offsets) {
-//                                final int target = this.index(offset);
-//                                worklist.put(target, new Variables(currentVars));
-//                            }
-//                            break;
-//                        } else if (var.isConditionalBranchInstruction()) {
-//                            final int jump = this.index(var.offset());
-//                            worklist.put(jump, new Variables(currentVars));
-//                            final int next = current + 1;
-//                            worklist.put(next, new Variables(currentVars));
-//                            break;
-//                        } else {
-//                            final int jump = this.index(var.offset());
-//                            worklist.put(jump, new Variables(currentVars));
-//                            break;
-//                        }
-//                    } else if (var.isReturnInstruction()) {
-//                        break;
-//                    }
-//                    if (var.isVarInstruction()) {
-//                        currentVars.put(var);
-//                    }
-//                }
-//                final Variables value = new Variables(currentVars);
-//                this.catches(current).stream().forEach(ind -> {
-//                    worklist.put(ind, new Variables(value));
-//                });
-//                all.put(current, value);
-//                current++;
-//            }
-//        }
-//        return all.values().stream().mapToInt(Variables::size).max().orElse(0);
     }
 
-    private List<Integer> catches(final int instruction) {
-        return this.tryblocks.stream().map(BytecodeTryCatchBlock.class::cast)
-            .filter(
-                block -> index(block.start()) <= instruction
-                    && index(block.end()) >= instruction
-            ).map(
-                block -> index(block.handler())
-            ).collect(Collectors.toList());
-    }
 
-    private int index(final Label label) {
-        for (int index = 0; index < this.instructions.size(); index++) {
-            final BytecodeEntry entry = this.instructions.get(index);
-            final BytecodeLabel obj = new BytecodeLabel(label, new AllLabels());
-            final boolean equals = entry.equals(obj);
-            if (equals) {
-                return index;
-            }
-        }
-        throw new IllegalStateException("Label not found");
-    }
-
-    @ToString
-    private static class Variables {
-
-        private final TreeMap<Integer, Integer> all;
-
-        public Variables() {
-            this(new HashMap<>(0));
-        }
-
-        public Variables(final Variables vars) {
-            this(vars.all);
-        }
-
-        public Variables(final Map<Integer, Integer> all) {
-            this.all = new TreeMap<>(all);
-        }
-
-        int size() {
-            if (all.isEmpty()) {
-                return 0;
-            }
-            final Map.Entry<Integer, Integer> entry = all.lastEntry();
-            return (entry.getKey() + 1) + ((int) (entry.getValue() * 0.5));
-        }
-
-        public void put(BytecodeInstruction var) {
-            this.put(var.localIndex(), var.size());
-        }
-
-        public void put(int index, int value) {
-            this.all.put(index, value);
-        }
-    }
 }
